requirements.txt
Flask
boto3
easyocr
Pillow
pdf2image
python-dotenv
